Moved game specific methods to new class
